fix: docker conflict error
$orchestration->run() sometimes also returns additional information
such as output from pulling docker images. As such, it's not reliable
to use the output for the container ID. Instead, we can just use the
container name for operations on the container.

Since we won't use the container ID, we can remove the 'id' column from
the activeRuntimes table.